requirements.txt
llama-cpp-python>=0.2.56
Pillow>=10.0.0
numpy>=1.24.0
ai-edge-litert
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.0.0
psutil>=5.9.0
python-multipart>=0.0.9
 
groq==0.9.0
sentence-transformers==3.0.1
fastapi==0.111.0
uvicorn==0.30.1
pymongo==4.7.2
motor==3.4.0          # async MongoDB driver
pydantic==2.7.1
python-dotenv==1.0.1
dotenv==0.9.9

fastapi==0.111.0
uvicorn==0.30.1
motor==3.4.0
pymongo==4.7.2
confluent-kafka==2.4.0
pydantic==2.7.1
python-dotenv==1.0.1
pytest==8.2.0
ruff==0.4.4